daily-news/requirements.txt
aiohttp==3.9.1
APScheduler==3.10.4
pyyaml==6.0.1
python-dotenv==1.0.0
tzlocal==5.2
# OCR dependencies - MinerU is preferred, PaddleOCR is fallback
magic-pdf[full]>=0.7.0
paddleocr==2.7.3
paddlepaddle==2.6.0
Pillow==10.1.0
